refactor(engineer_loop/execution): split tests to new sibling tests_execution_extra.rs (#1266)
Existing tests_execution.rs (336 LOC) was already a sibling; created NEW
tests_execution_extra.rs for the inline-mod tests.

- src/engineer_loop/execution.rs: 669 → 412 LOC
- New src/engineer_loop/tests_execution_extra.rs (258 LOC)

Pure refactor: 20 tests pass.
